Refactor and add backend api for checks reorder
